refactor workflow and storage types in steveo package
